Reorganization and renaming of some functions and files in prep for adding giphy api call
WeirdnessSelect component will need some reworking, as it's using the onSubmit to show strings. Will need to rewrite the component to submit the weirdness and search term to the giphy api to return a single gif. I'll then rewrite the current add_gif functionality to function similar to the current iteration to then pass the gif's url and weirdness to the LikedGifs component.