Fixed giphy api issue after some tinkering.
Took some time to figure out why the giphy api was returning html opposed to the data, after some tinkering, I just ended up pulling the promise out of fetch and manually returning it to the later .then.

Image is still not showing up on the page, but that is the next thing to work on.