Users can now press the Like button to like the current gif.
When the user presses the like button, it's sent to the LikedGifs component, which lists the gifs they've liked and will eventually calculate the score.

Worked a little bit on CSS and getting the gifs sized. Still needs some work in this department, but I can style after the functionality is in place.

Next task is to allow the user to unlike a gif from their list.