Added functionality to prevent user from liking the same gif multiple times.
I had to do this, because I was getting issues with my removal code where all my gifs had the same id... Not sure how uuid is creating the same id over and over, but I figure it's something to do with the way the code is compiled... Somewhat akin to a seed in a video game or random number generator.

Either way, I decided to go with this route because it was the quickest to implement. I'd probably debug the main issue instead of just not allowing the user to select the same gif twice, but again due to time constrants, I'm choosing this method.

Next task is to finish the removal code so the user can remove a gif from their list.