Implimented code to prevent users from selecting multiple gifs from the same search term
This one required a bit of a rewrite. I had to add the search term as a property of the shown gif. It took me some time to realize that this was what was needed to be done, as I was having issues. After giving the shown gif the term that was used to find it, I was quickly able to add some code to prevent users from selecting not only duplicate gifs, but also gifs within the same search term.

Next task is to FINALLY start fixing the remove gif button from the liked gifs now that all my gifs are unique and I have no worries about dealing with duplicates.