Some more css fixes
Decided to finish the day off with some css fixes.

The loading circle now is properly shown near the center of the left component while waiting for the API call.

Added CSS grids and some text changes to format the right side to be a bit cleaner.

Moved the Calculate Score button into its own gridded div so that it is not pushed down by gifs if the user selects  more than 5. Will limit gif selection to 5 tomorrow.