Gif Unliking has been implemented and working properly
This was much easier to implement, as most of the framework in the actions/reducer had already been implemented. Just took some tweaking to get it working properly and a forceUpload() in LikedGifs to re-render the component. I'm not too sure why the component was not re-rendering, but that could be something to look at later if I have time.

Went ahead and cleaned up all my comments, whitespace and unneeded code to make the files more readable by others.

Next Task is to implement the calculate score button to return a new route with the five liked gifs and the user's weirdness score.