CalculateScore Component functionality is working properly
The CalculateScore component is now properly posting the five liked gifs with their information (title/weirdness/search term), and then calculates the user's weirdness based off those properties.

The Play Again button properly brings the user back to the home page to play again, but needs to have a separate redux dispatch action to clear the store so the user may start from a fresh slate.

Added five gif limit to liking gifs. User will not be able to like more gifs if they've already liked five, and are prompted to either unlike a gif to choose another or calculate their score.

Next Task is to add the functionality to clear the redux store to a blank slate so the user may replay.