WeirdnessSelect component now re-initializes the redux store to the initial state when mounted
I tried a couple different approaches to resetting the redux store to the initial state and felt that this method was the most appropriate for the application. The Redux store will be reset to it's initial state every time WeirdnessSelect is mounted, which only occurs on the initial launch of the app, and whenever the user clicks "Play Again" on the CalculateScore page when redirected back to the "/" route.

As of this commit, all functionality required for the project assignment has been completed.

There are some minor things such as saving the value of the slider between likes and some css fixes that need to be worked on prior to completion.