RangeSlider value is updated to the previously liked gif's weirdness value
I decided to do it this way, because it allows the user to keep the weirdness number that they used on the previous gif, which makes it a bit more intuative to move it up or down from that point for the next gif.

At this point in the project, it's functionally complete. Everything works as expected and I am unable to break the application myself.

I will add some minor CSS changes such as styled buttons and some user feedback, but beyond that, I don't really want to deviate from the Mockup provided.

Last Task is to add some css styling to give the user some more tactile feedback.