Fixed small bug and added css to buttons
Added some small css styling to the buttons to give them a bit of feedback to the user.

I also found a bug that allowed the user to click like without clicking search. I've removed the user's ability to do that and can only press the like button when a gif is active.