Fixed some bugs
I found a bug in which the giphy gif imported had no title associated with it, and it occured very rarely. I now set the gif's search term to the title if the gif contains no title.

Fixed a bug with the range slider, where after the user calculated their score and went back to the home page, the slider was visibly showing the previous showngif's weirdness value while still having a value of 0. This resulted in the user seeing a value of 10, but when searching, would have the search use a value of 0 resulting in the wrong search.